+++ Sources/ApiKit/ApiEnvironment.swift
@@ -21,8 +21,24 @@ import Foundation
 /// headers and query parameters they need. Environments can
 /// define global headers and query parameters, while routes
 /// can define route-specific ones.
-public protocol ApiEnvironment: ApiRequestData, Sendable {
+public protocol ApiEnvironment: Sendable {
+    
+    /// Optional header parameters to apply to all requests.
+    var headers: [String: String]? { get }
+
+    /// Optional query params to apply to all requests.
+    var queryParams: [String: String]? { get }
 
     /// The base URL of the environment.
     var url: String { get }
+}
+
+extension ApiEnvironment {
+
+    /// Convert ``queryParams`` to url encoded query items.
+    var encodedQueryItems: [URLQueryItem]? {
+        queryParams?
+            .map { URLQueryItem(name: $0.key, value: $0.value) }
+            .sorted { $0.name < $1.name }
+    }
 }